+++ src/frontend/komponenter/Infotrygd/Vedtakstabell.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 
 import { Normaltekst } from 'nav-frontend-typografi';
 
-import { IInfotrygdStønad } from '../../typer/infotrygd';
+import { IInfotrygdSak, IInfotrygdStønad } from '../../typer/infotrygd';
 
 const IngenVedtakTekst = styled(Normaltekst)`
     margin: 1rem;
@@ -14,7 +14,7 @@ const antallBarn = (stønad: IInfotrygdStønad) => {
     return stønad.barn.length;
 };
 
-export const Vedtakstabell: React.FC<{ stønader: IInfotrygdStønad[] }> = ({ stønader }) => {
+export const Vedtakstabell: React.FC<{ saker: IInfotrygdSak[] }> = ({ saker }) => {
     return (
         <>
             <table className="tabell">
@@ -33,25 +33,25 @@ export const Vedtakstabell: React.FC<{ stønader: IInfotrygdStønad[] }> = ({ st
                     </tr>
                 </thead>
                 <tbody>
-                    {stønader.map((stønad: IInfotrygdStønad, index: number) => {
+                    {saker.map((sak: IInfotrygdSak, index: number) => {
                         return (
                             <tr key={index}>
-                                <td>{stønad.iverksattFom}</td>
-                                <td>{stønad.virkningFom}</td>
+                                <td>{sak.stønad ? sak.stønad.iverksattFom : ''}</td>
+                                <td>{sak.stønad ? sak.stønad.virkningFom : ''}</td>
                                 <td>{/* kommer når vi finner dataene i replikasettet */}</td>
-                                <td>{stønad.status}</td>
-                                <td>{stønad.tekstkode}</td>
-                                <td>{antallBarn(stønad)}</td>
-                                <td>{stønad.opphørtIver}</td>
-                                <td>{stønad.opphørtFom}</td>
-                                <td>{stønad.opphørsgrunn}</td>
+                                <td>{sak.stønad ? sak.stønad.status : ''}</td>
+                                <td>{sak.stønad ? sak.stønad.tekstkode : ''}</td>
+                                <td>{sak.stønad ? antallBarn(sak.stønad) : ''}</td>
+                                <td>{sak.stønad ? sak.stønad.opphørtIver : ''}</td>
+                                <td>{sak.stønad ? sak.stønad.opphørtFom : ''}</td>
+                                <td>{sak.stønad ? sak.stønad.opphørsgrunn : ''}</td>
                                 <td>{/* detaljer, kommer senere */}</td>
                             </tr>
                         );
                     })}
                 </tbody>
             </table>
-            {stønader.length === 0 ? <IngenVedtakTekst>Ingen vedtak.</IngenVedtakTekst> : undefined}
+            {saker.length === 0 ? <IngenVedtakTekst>Ingen vedtak.</IngenVedtakTekst> : undefined}
         </>
     );
 };
